fix: drop redundant PortNumber/int32 conversions (unconvert)
gateway-api v1.6.0 defines PortNumber as an int32 alias, so the explicit
int32()/PortNumber() conversions are no-ops that fail the unconvert linter
after master merged the Gateway API 1.6 upgrade.

+++ internal/adc/translator/tcproute.go
@@ -50,7 +50,7 @@ func newDefaultUpstreamWithoutScheme() *adctypes.Upstream {
 func listenerPortSet(tctx *provider.TranslateContext) map[int32]struct{} {
 	portSet := make(map[int32]struct{}, len(tctx.Listeners))
 	for _, listener := range tctx.Listeners {
-		portSet[int32(listener.Port)] = struct{}{}
+		portSet[listener.Port] = struct{}{}
 	}
 	return portSet
 }
